refactor(client): extract the Entities tab into a self-loading OnPush component (A17.9b-6e)
Second record tab out of the shell, following the memories pattern.
entities-tab.component owns the entity create form, inline edit, delete, and its own
entity-search / type-tag filter / pagination + loader. Self-loads via a spaceId effect;
emits `mutated` so the shell refreshes tab-count stats; the shell's loadCurrentTab now
early-returns for entities.

Entity-specific behaviour preserved (pinned by the relocated 6b case): both create AND
inline-edit strip empty optional properties via the entity schema (unlike memory, which
sends them raw); entity search uses the <app-entity-search> bar (semantic default) with
no per-tab search-mode pill. Reuses the shared brain-table.styles.ts.

The 6b createEntity characterization case moved to entities-tab.component.spec.ts
alongside new self-load / edit / delete / search / pagination tests.

brain.component.ts 2070 -> 1762. Client suite 190 -> 197. tsc --noUnusedLocals clean;
prod AOT build warning-free bar the pre-existing qrcode CJS bailout.
