+++ client/src/app/pages/brain/brain.component.ts
@@ -7,6 +7,7 @@ import { RecordDrawerComponent } from './record-drawer.component';
 import { QueryTabComponent } from './query-tab.component';
 import { RecordListState } from './record-list-state.service';
 import { MemoriesTabComponent } from './memories-tab.component';
+import { EntitiesTabComponent } from './entities-tab.component';
 import { BRAIN_CHIP_STYLES } from './brain-form.styles';
 import { toLocalDatetime, fmtApiError } from './brain-format';
 import { FormsModule } from '@angular/forms';
@@ -47,7 +48,7 @@ interface SpaceView {
   // or happens in a template event handler, both of which mark the view dirty. That coupling is
   // load-bearing and pinned by the specs (the drawer's own version lives in the drawer component).
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [CommonModule, FormsModule, GraphComponent, FileManagerComponent, EntitySearchComponent, PropertiesViewComponent, PropertiesEditorComponent, TagInputComponent, PhIconComponent, ErrorStateComponent, RecordFilterBarComponent, RecordDrawerComponent, QueryTabComponent, MemoriesTabComponent, TranslocoPipe],
+  imports: [CommonModule, FormsModule, GraphComponent, FileManagerComponent, EntitySearchComponent, PropertiesViewComponent, PropertiesEditorComponent, TagInputComponent, PhIconComponent, ErrorStateComponent, RecordFilterBarComponent, RecordDrawerComponent, QueryTabComponent, MemoriesTabComponent, EntitiesTabComponent, TranslocoPipe],
   providers: [BrainStore, EntityRefPicker, RecordDrawerState, RecordListState],
   styles: [BRAIN_CHIP_STYLES, `
     .space-tabs {
@@ -467,180 +468,7 @@ interface SpaceView {
         @if (activeTab() === 'memories') { <app-memories-tab [spaceId]="activeSpaceId()" (mutated)="loadStats(activeSpaceId())" /> }
 
         <!-- Entities -->
-        @if (activeTab() === 'entities') {
-
-          <div class="content-header">
-            <app-entity-search
-              mode="bar"
-              [spaceId]="activeSpaceId()"
-              placeholder="common.searchEntitiesPlaceholder"
-              defaultMode="semantic"
-              (queryChange)="onEntitySearchChange($event)"
-              (cleared)="onEntitySearchClear()"
-              (selected)="onEntitySearchPick($event)"
-            />
-            <button class="btn-primary btn btn-sm" (click)="openEntityForm()" [disabled]="showEntityForm()">{{ 'brain.entities.addButton' | transloco }}</button>
-          </div>
-          <div class="list-filter-row">
-            <app-record-filter-bar
-              [typeOptions]="store.entityTypeOptions()"
-              [tagSuggestions]="store.entityTagSuggestions()"
-              [value]="recordFilter()"
-              (filterChange)="onFilterChange($event)"
-            />
-          </div>
-
-          @if (showEntityForm()) {
-            <form class="create-form" (ngSubmit)="createEntity()">
-              <div class="field" style="flex:1; min-width:140px;">
-                <label>{{ 'brain.entities.table.name' | transloco }}</label>
-                <input type="text" [(ngModel)]="entityForm.name" name="name" required />
-              </div>
-              <div class="field" style="width:140px;">
-                <label>Type @if (store.entityTypeNames().length) { <span style="color:var(--error)">*</span> }</label>
-                @if (store.entityTypeNames().length) {
-                  <select [(ngModel)]="entityForm.type" name="type" required (ngModelChange)="onEntityTypeChange($event, 'create')">
-                    @for (t of store.entityTypeNames(); track t) {
-                      <option [value]="t">{{ t }}</option>
-                    }
-                  </select>
-                } @else {
-                  <input type="text" [(ngModel)]="entityForm.type" name="type" [placeholder]="'brain.entities.form.typePlaceholder' | transloco" />
-                }
-              </div>
-              <div class="field" style="flex:1; min-width:180px;">
-                <label>{{ 'brain.entities.table.tags' | transloco }}</label>
-                <app-tag-input [(value)]="entityForm.tags" [suggestions]="store.entityTagSuggestions()" inputName="entFormTags" />
-              </div>
-              <div class="field" style="flex:1; min-width:200px;">
-                <label>{{ 'brain.entities.table.description' | transloco }}</label>
-                <textarea [(ngModel)]="entityForm.description" name="description" rows="3" style="resize:vertical;"></textarea>
-              </div>
-              <div class="field" style="flex:1; min-width:220px;">
-                <label>{{ 'brain.entities.table.properties' | transloco }}</label>
-                <app-properties-editor
-                  [schema]="store.entitySchema(entityForm.type)"
-                  [required]="store.requiredProps(store.entitySchema(entityForm.type))"
-                  [(value)]="entityForm.properties"
-                />
-              </div>
-              <button class="btn-primary btn btn-sm" type="submit" [disabled]="creatingEntity() || !entityForm.name.trim() || (store.entityTypeNames().length ? !entityForm.type : false)">
-                @if (creatingEntity()) { <span class="spinner" style="width:12px;height:12px;border-width:2px;"></span> }
-                {{ 'common.save' | transloco }}
-              </button>
-              <button class="btn-secondary btn btn-sm" type="button" (click)="showEntityForm.set(false)">{{ 'common.cancel' | transloco }}</button>
-            </form>
-          }
-
-          @if (createEntityError()) {
-            <div class="alert alert-error" style="margin-bottom:12px;">{{ createEntityError() }}</div>
-          }
-
-          <div class="table-wrapper">
-            <table>
-              <thead>
-                <tr>
-                  <th>{{ 'brain.entities.table.name' | transloco }}</th><th>{{ 'brain.entities.table.type' | transloco }}</th><th>{{ 'brain.entities.table.description' | transloco }}</th><th>{{ 'brain.entities.table.tags' | transloco }}</th><th>{{ 'brain.entities.table.properties' | transloco }}</th><th>{{ 'brain.entities.table.created' | transloco }}</th><th></th>
-                </tr>
-              </thead>
-              <tbody>
-                @for (ent of store.entities(); track ent._id) {
-                  @if (recordList.editingId() === ent._id) {
-                    <tr>
-                      <td colspan="7">
-                        <div class="create-form" style="border:none; padding:8px 0;">
-                          <div class="field" style="flex:1; min-width:120px; margin-bottom:0;">
-                            <label>{{ 'brain.entities.table.name' | transloco }}</label>
-                            <input type="text" [(ngModel)]="editEntity.name" name="editEntName" />
-                          </div>
-                          <div class="field" style="width:120px; margin-bottom:0;">
-                            <label>Type @if (store.entityTypeNames().length) { <span style="color:var(--error)">*</span> }</label>
-                            @if (store.entityTypeNames().length) {
-                              <select [(ngModel)]="editEntity.type" name="editEntType" (ngModelChange)="onEntityTypeChange($event, 'inline')">
-                                @for (t of store.entityTypeNames(); track t) {
-                                  <option [value]="t">{{ t }}</option>
-                                }
-                              </select>
-                            } @else {
-                              <input type="text" [(ngModel)]="editEntity.type" name="editEntType" />
-                            }
-                          </div>
-                          <div class="field" style="flex:1; min-width:160px; margin-bottom:0;">
-                            <label>{{ 'brain.entities.table.description' | transloco }}</label>
-                            <textarea [(ngModel)]="editEntity.description" name="editEntDesc" rows="2" style="resize:vertical;"></textarea>
-                          </div>
-                          <div class="field" style="flex:1; min-width:180px; margin-bottom:0;">
-                            <label>{{ 'brain.entities.table.tags' | transloco }}</label>
-                            <app-tag-input [(value)]="editEntity.tags" [suggestions]="store.entityTagSuggestions()" inputName="entEditTags" />
-                          </div>
-                          <div class="field" style="flex:1; min-width:220px; margin-bottom:0;">
-                            <label>{{ 'brain.entities.table.properties' | transloco }}</label>
-                            <app-properties-editor
-                              [schema]="store.entitySchema(editEntity.type)"
-                              [required]="store.requiredProps(store.entitySchema(editEntity.type))"
-                              [(value)]="editEntity.properties"
-                            />
-                          </div>
-                          <div style="display:flex; gap:6px; align-items:flex-end;">
-                            <button class="btn btn-sm btn-primary" [disabled]="recordList.editSaving()" (click)="saveEditEntity(ent._id)">
-                              @if (recordList.editSaving()) { <span class="spinner" style="width:11px;height:11px;border-width:2px;"></span> } Save
-                            </button>
-                            <button class="btn btn-sm btn-secondary" (click)="recordList.cancelEdit()">{{ 'common.cancel' | transloco }}</button>
-                          </div>
-                          @if (recordList.editError()) { <div style="font-size:12px; color:var(--error);">{{ recordList.editError() }}</div> }
-                        </div>
-                      </td>
-                    </tr>
-                  } @else {
-                    <tr>
-                      <td>{{ ent.name }}</td>
-                      <td>
-                        @if (ent.type) { <span class="badge badge-purple">{{ ent.type }}</span> }
-                      </td>
-                      <td class="desc-cell" style="max-width:200px;" [title]="ent.description ?? ''">
-                        {{ ent.description || '—' }}
-                      </td>
-                      <td style="font-size:11px;">
-                        @for (tag of (ent.tags ?? []); track tag) { <span class="tag">{{ tag }}</span> }
-                        @if (!(ent.tags?.length)) { <span style="color:var(--text-muted)">—</span> }
-                      </td>
-                      <td><app-properties-view [properties]="ent.properties" [schema]="store.entitySchema(ent.type)" /></td>
-                      <td style="color:var(--text-muted)">{{ ent.createdAt | date:'dd.MM.yyyy' }}</td>
-                      <td style="white-space:nowrap;">
-                        <button class="icon-btn" [attr.title]="'common.viewDetails' | transloco" [attr.aria-label]="'common.viewDetails' | transloco" (click)="drawerState.open('entity', ent)"><ph-icon name="eye" [size]="16"/></button>
-                        @if (recordList.confirmDeleteId() === ent._id) {
-                          <span class="inline-confirm">
-                            Delete?
-                            <button class="btn btn-sm btn-danger" (click)="deleteEntity(ent._id)">{{ 'common.yes' | transloco }}</button>
-                            <button class="btn btn-sm btn-secondary" (click)="cancelDelete()">{{ 'common.no' | transloco }}</button>
-                          </span>
-                        } @else {
-                          <button class="icon-btn danger" [attr.aria-label]="'brain.entities.deleteAriaLabel' | transloco" (click)="requestDelete(ent._id)"><ph-icon name="x" [size]="16"/></button>
-                        }
-                      </td>
-                    </tr>
-                  }
-                } @empty {
-                  <tr><td colspan="7">
-                    @if (recordList.loadError() !== null) {
-                      <app-error-state [message]="'brain.error.loadEntities' | transloco" [reason]="recordList.loadError() ?? ''" (retry)="retryCurrentTab()" />
-                    } @else {
-                    <div class="empty-state" style="padding:32px">
-                      <div class="empty-state-icon"><ph-icon name="tag" [size]="48"/></div>
-                      <h3>{{ 'brain.entities.empty.title' | transloco }}</h3>
-                    </div>
-                    }
-                  </td></tr>
-                }
-              </tbody>
-            </table>
-          </div>
-          <div class="pagination">
-            <button class="btn btn-sm btn-secondary" [disabled]="entitySkip() === 0" (click)="prevEntityPage()"><ph-icon name="arrow-left" [size]="14" style="display:inline-flex;vertical-align:middle;"/> {{ 'common.prev' | transloco }}</button>
-            <span class="pager-info">{{ store.entities().length ? (entitySkip() + 1) + '–' + (entitySkip() + store.entities().length) : '–' }}</span>
-            <button class="btn btn-sm btn-secondary" [disabled]="store.entities().length < pageSize" (click)="nextEntityPage()">{{ 'common.next' | transloco }} <ph-icon name="arrow-right" [size]="14" style="display:inline-flex;vertical-align:middle;"/></button>
-          </div>
-        }
+        @if (activeTab() === 'entities') { <app-entities-tab [spaceId]="activeSpaceId()" (mutated)="loadStats(activeSpaceId())" /> }
 
         <!-- Edges -->
         @if (activeTab() === 'edges') {
@@ -1343,7 +1171,6 @@ export class BrainComponent implements OnInit {
     this.recordFilter.set(f);
     switch (this.activeTab()) {
       case 'memories': this.skip.set(0); break;
-      case 'entities': this.entitySkip.set(0); break;
       case 'edges': this.edgeSkip.set(0); break;
       case 'chrono': this.chronoSkip.set(0); break;
       default: break;
@@ -1355,9 +1182,6 @@ export class BrainComponent implements OnInit {
   skip = signal(0);
   filterEntity = signal('');
 
-  // Entities pagination + search
-  entitySkip = signal(0);
-  entitySearch = signal('');
   private _edgeSemTimer: ReturnType<typeof setTimeout> | null = null;
   private _chronoSemTimer: ReturnType<typeof setTimeout> | null = null;
 
@@ -1372,15 +1196,8 @@ export class BrainComponent implements OnInit {
   reindexing = signal(false);
   reindexResult = signal('');
 
-  editEntity = { name: '', type: '', tags: [] as string[], description: '', properties: {} as Record<string, string | number | boolean> };
   editEdge = { from: '', to: '', fromName: undefined as string | undefined, toName: undefined as string | undefined, label: '', weight: null as number | null, tags: [] as string[], description: '', properties: {} as Record<string, string | number | boolean> };
   editChrono = { title: '', kind: '' as string, status: '' as string, startsAt: '', endsAt: '', description: '', tags: [] as string[], entityIds: '' };
-
-  // Create entity form
-  showEntityForm = signal(false);
-  creatingEntity = signal(false);
-  createEntityError = signal('');
-  entityForm = { name: '', type: '', tags: [] as string[], description: '', properties: {} as Record<string, string | number | boolean> };
 
   // Create edge form
   showEdgeForm = signal(false);
@@ -1434,12 +1251,10 @@ export class BrainComponent implements OnInit {
     this.picker.spaceId.set(id);
     this.drawerState.spaceId.set(id);
     this.skip.set(0);
-    this.entitySkip.set(0);
     this.edgeSkip.set(0);
     this.chronoSkip.set(0);
     this.recordFilter.set({ type: '', tag: '' });
     this.filterEntity.set('');
-    this.entitySearch.set('');
     this.store.memorySearch.set('');
     this.store.edgeSearch.set('');
     this.store.chronoSearch.set('');
@@ -1456,7 +1271,6 @@ export class BrainComponent implements OnInit {
   setTab(tab: BrainTab): void {
     this.activeTab.set(tab);
     this.skip.set(0);
-    this.entitySkip.set(0);
     this.edgeSkip.set(0);
     this.chronoSkip.set(0);
     this.fileMetaSkip.set(0);
@@ -1473,9 +1287,6 @@ export class BrainComponent implements OnInit {
     this.loadCurrentTab(this.activeSpaceId());
   }
 
-  prevEntityPage(): void { this.entitySkip.update(s => Math.max(0, s - this.pageSize)); this.loadCurrentTab(this.activeSpaceId()); }
-  nextEntityPage(): void { this.entitySkip.update(s => s + this.pageSize); this.loadCurrentTab(this.activeSpaceId()); }
-
   prevEdgePage(): void { this.edgeSkip.update(s => Math.max(0, s - this.pageSize)); this.loadCurrentTab(this.activeSpaceId()); }
   nextEdgePage(): void { this.edgeSkip.update(s => s + this.pageSize); this.loadCurrentTab(this.activeSpaceId()); }
 
@@ -1488,38 +1299,6 @@ export class BrainComponent implements OnInit {
   onFileMetaSearch(q: string): void {
     this.store.fileMetaSearch.set(q);
     // client-side filter via filteredFileMetas computed() — no API call per keystroke
-  }
-
-  searchEntities(): void { this.entitySkip.set(0); this.loadCurrentTab(this.activeSpaceId()); }
-
-  // ── Entity search bar handlers (brain entities tab) ──────────────────────
-  onEntitySearchChange(q: string): void {
-    this.entitySearch.set(q);
-    this.entitySkip.set(0);
-    this.loadEntitiesSilent();
-  }
-  onEntitySearchClear(): void {
-    this.entitySearch.set('');
-    this.entitySkip.set(0);
-    this.loadEntitiesSilent();
-  }
-  onEntitySearchPick(ent: Entity): void {
-    this.entitySearch.set(ent.name);
-    this.entitySkip.set(0);
-    this.loadEntitiesSilent();
-  }
-
-  loadEntitiesSilent(): void {
-    const spaceId = this.activeSpaceId();
-    if (!spaceId) return;
-    const ef: { search?: string; type?: string; tag?: string } = {};
-    if (this.entitySearch()) ef.search = this.entitySearch();
-    if (this.recordFilter().type) ef.type = this.recordFilter().type;
-    if (this.recordFilter().tag) ef.tag = this.recordFilter().tag;
-    this.brainApi.listEntities(spaceId, this.pageSize, this.entitySkip(), ef).subscribe({
-      next: ({ entities }) => this.store.entities.set(entities),
-      error: () => {},
-    });
   }
 
   onEdgeSearch(q: string): void {
@@ -1630,21 +1409,11 @@ export class BrainComponent implements OnInit {
   private loadCurrentTab(spaceId: string): void {
     if (!spaceId) return;
     if (this.activeTab() === 'memories') return; // self-loading component
+    if (this.activeTab() === 'entities') return; // self-loading component
     this.recordList.loading.set(true);
     this.recordList.loadError.set(null);
 
     switch (this.activeTab()) {
-      case 'entities': {
-        const ef: { search?: string; type?: string; tag?: string } = {};
-        if (this.entitySearch()) ef.search = this.entitySearch();
-        if (this.recordFilter().type) ef.type = this.recordFilter().type;
-        if (this.recordFilter().tag) ef.tag = this.recordFilter().tag;
-        this.brainApi.listEntities(spaceId, this.pageSize, this.entitySkip(), ef).subscribe({
-          next: ({ entities }) => { this.store.entities.set(entities); this.recordList.loading.set(false); },
-          error: (e) => { this.recordList.loadError.set(httpErrorReason(e)); this.recordList.loading.set(false); },
-        });
-        break;
-      }
       case 'edges': {
         const gf: { type?: string; tag?: string } = {};
         if (this.recordFilter().type) gf.type = this.recordFilter().type;
@@ -1723,18 +1492,6 @@ export class BrainComponent implements OnInit {
   requestDelete(id: string): void { this.recordList.confirmDeleteId.set(id); }
   cancelDelete(): void { this.recordList.confirmDeleteId.set(''); }
 
-  startEditEntity(ent: Entity): void {
-    this.recordList.editingId.set(ent._id);
-    this.recordList.editError.set('');
-    this.editEntity = {
-      name: ent.name,
-      type: ent.type ?? '',
-      tags: ent.tags ?? [],
-      description: ent.description ?? '',
-      properties: this.store.buildPropertiesObject('entity', ent.properties ?? {}, ent.type),
-    };
-  }
-
   startEditEdge(edge: Edge): void {
     this.recordList.editingId.set(edge._id);
     this.recordList.editError.set('');
@@ -1764,26 +1521,6 @@ export class BrainComponent implements OnInit {
       tags: entry.tags ?? [],
       entityIds: (entry.entityIds ?? []).join(', '),
     };
-  }
-
-  saveEditEntity(id: string): void {
-    this.recordList.editSaving.set(true);
-    this.recordList.editError.set('');
-    const entProps = this.store.stripEmptyOptionalProps(this.editEntity.properties, this.store.entitySchema(this.editEntity.type));
-    this.brainApi.updateEntity(this.activeSpaceId(), id, {
-      name: this.editEntity.name.trim(),
-      type: this.editEntity.type.trim(),
-      tags: this.editEntity.tags,
-      description: this.editEntity.description.trim(),
-      ...(Object.keys(entProps).length ? { properties: entProps } : {}),
-    }).subscribe({
-      next: (updated) => {
-        this.recordList.editSaving.set(false);
-        this.recordList.editingId.set('');
-        this.store.entities.update(list => list.map(e => e._id === id ? updated : e));
-      },
-      error: (err) => { this.recordList.editSaving.set(false); this.recordList.editError.set(fmtApiError(err, 'Failed to save')); },
-    });
   }
 
   saveEditEdge(id: string): void {
@@ -1897,28 +1634,6 @@ export class BrainComponent implements OnInit {
     this.setTab('files');
   }
 
-  createEntity(): void {
-    if (!this.entityForm.name.trim()) return;
-    this.creatingEntity.set(true);
-    this.createEntityError.set('');
-    const body: Parameters<BrainApi['createEntity']>[1] = { name: this.entityForm.name.trim() };
-    if (this.entityForm.type.trim()) body.type = this.entityForm.type.trim();
-    if (this.entityForm.tags.length) body.tags = this.entityForm.tags;
-    if (this.entityForm.description.trim()) body.description = this.entityForm.description.trim();
-    const props = this.store.stripEmptyOptionalProps(this.entityForm.properties, this.store.entitySchema(this.entityForm.type));
-    if (Object.keys(props).length) body.properties = props;
-    this.brainApi.createEntity(this.activeSpaceId(), body).subscribe({
-      next: () => {
-        this.creatingEntity.set(false);
-        this.showEntityForm.set(false);
-        this.entityForm = { name: '', type: '', tags: [], description: '', properties: {} as Record<string, string | number | boolean> };
-        this.loadStats(this.activeSpaceId());
-        this.loadCurrentTab(this.activeSpaceId());
-      },
-      error: (err) => { this.creatingEntity.set(false); this.createEntityError.set(fmtApiError(err, 'Failed to create entity')); },
-    });
-  }
-
   createEdge(): void {
     if (!this.edgeForm.from.trim() || !this.edgeForm.to.trim() || !this.edgeForm.label.trim()) return;
     this.creatingEdge.set(true);
@@ -1942,14 +1657,6 @@ export class BrainComponent implements OnInit {
         this.loadCurrentTab(this.activeSpaceId());
       },
       error: (err) => { this.creatingEdge.set(false); this.createEdgeError.set(fmtApiError(err, 'Failed to create edge')); },
-    });
-  }
-
-  deleteEntity(id: string): void {
-    this.recordList.confirmDeleteId.set('');
-    this.brainApi.deleteEntity(this.activeSpaceId(), id).subscribe({
-      next: () => { this.store.entities.update(list => list.filter(e => e._id !== id)); this.loadStats(this.activeSpaceId()); },
-      error: () => {},
     });
   }
 
@@ -2025,21 +1732,6 @@ export class BrainComponent implements OnInit {
   }
 
   // ── Form openers with schema prefill ────────────────────────────────────
-
-  openEntityForm(): void {
-    const firstType = Object.keys(this.store.spaceMeta()?.typeSchemas?.entity ?? {})[0] ?? '';
-    this.entityForm = { name: '', type: firstType, tags: [], description: '', properties: this.store.buildPropertiesObject('entity', {}, firstType) };
-    this.showEntityForm.set(true);
-  }
-
-  /** Called when the entity type dropdown changes. Rebuilds properties: keeps existing values, adds defaults for any new schema-required fields. */
-  onEntityTypeChange(type: string, target: 'create' | 'inline'): void {
-    if (target === 'create') {
-      this.entityForm.properties = this.store.buildPropertiesObject('entity', this.entityForm.properties, type);
-    } else {
-      this.editEntity.properties = this.store.buildPropertiesObject('entity', this.editEntity.properties, type);
-    }
-  }
 
   openEdgeForm(): void {
     const firstLabel = Object.keys(this.store.spaceMeta()?.typeSchemas?.edge ?? {})[0] ?? '';
